# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mrprototcol",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "https://mrprotocol.com/api/v1",
            "description": "Base URL"
        },
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Local Base URL"
        }
    ],
    "paths": {
...
        "/auth/reset-password": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Change user password after email verification",
                "description": "Change user password after email verification.",
                "operationId": "b773aaaeaa6d02a1785bb47b8d05fb91",
                "requestBody": {
                    "description": "Request body containing the verification token and new password",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Verification token received in the email",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "New password",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "Confirmation of the new password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful password change",
                                            "type": "string",
                                            "example": "Password changed successfully"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating expired verification token",
                                            "type": "string",
                                            "example": "Verification token has expired"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 400
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating invalid verification token",
                                            "type": "string",
                                            "example": "Invalid verification token"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "error"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "integer",
                                            "example": 404
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/auth/google/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Log in with Google authentication",
                "description": "Log in with Google authentication.",
                "operationId": "9f4603884c7ea31616b823f550b12f04",
                "requestBody": {
                    "description": "Request body containing authentication code",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "authCode"
                                ],
                                "properties": {
                                    "authCode": {
                                        "description": "Authentication code received from Google",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "description": "Message indicating successful login",
                                            "type": "string",
                                            "example": "Login successful"
                                        },
                                        "status": {
                                            "description": "Status of the response",
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "statusCode": {
                                            "description": "HTTP status code",
                                            "type": "string",
                                            "example": "200"
                                        },
                                        "accessToken": {
                                            "description": "Access token for the user",
                                            "type": "string",
                                            "format": "text"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/UserResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "UserResource": {
                "title": "User Resource",
                "description": "User resource representation",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "John Doe"
                    },
                    "email": {
                        "type": "string",
                        "example": "john@example.com"
                    },
                    "roles": {
                        "type": "array",
                        "items": {
                            "type": "string",
                            "example": "user"
                        }
                    }
                },
                "type": "object"
            }
        },
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}